pyproject.toml
[project]
name = "photo-renamer"
version = "0.1.0"
description = "Uses a vision model to generate descriptive, unique filenames for photos."
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "rich>=14.3.3",
    "pillow>=10.2.0",
    "local-first-common",
]

[project.scripts]
rename-photo = "photo_renamer.logic:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_renamer"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv.sources]
local-first-common = { git = "https://github.com/jamalhansen/local-first-common.git", branch = "main" }

[tool.pytest.ini_options]
pythonpath = ["src"]
